backend-updates/requirements.txt
requests
python-dotenv
bs4
supabase
apscheduler